Mathematica notebook
This Mathematica notebook includes the computations to generate data and draw figures in the following papers.
[1] M. Koiso and U. Miyamoto, JSIAM Letters 15, pp.9-12 (2023) [arXiv:2306.12211]
[2] M. Koiso and U. Miyamoto, to appear in JJIAM (2023) [arXiv:1905.01705]
where [1] is a letter version of full paper [2]. If one uses or arranges this notebook in your analysis for any publication, he/she has to cite the above papers appropriately. In this notebook, we calculate the mean curvature H, surface area A, and volume V of spheres, cylinders, and unduloids in R^(n+2), and then draw (V, A) diagrams for several n's. In addition, we draw (s, H'(s)) and (s, V'(s)) diagrams, where s is the parameter characterizing the family of unduloids.